Updated version to 1.250.
Added full multi-threaded download status updating to the class without the need for you, the programmer, to write any multi-threaded code beyond creating the thread to download the file in. The class can handle all multi-threaded download status updating for you including calculating how much data is downloaded per second.

All you have to do, as the programmer, is to inject your download status updating code in as a routine into the class instance as a plugged-in code or subroutine and the class will do all the work for you.